Fixed clock arrow moving smoothly
