+++ Cycle/UI/ClockUI.tscn
@@ -1,9 +1,9 @@
-[gd_scene load_steps=4 format=3 uid="uid://cfuw3qet6mhnq"]
+[gd_scene load_steps=5 format=3 uid="uid://cfuw3qet6mhnq"]
 
+[ext_resource type="Script" path="res://Cycle/DayNight/ClockUi.cs" id="1_jcn3g"]
 [ext_resource type="Texture2D" uid="uid://boj71d2pygk12" path="res://Cycle/UI/daynightcycle-ui.png" id="1_knyeu"]
 [ext_resource type="FontFile" uid="uid://dt6jbuskqvs7g" path="res://Cycle/UI/fonts/pixelmix/pixelmix.ttf" id="2_piu5q"]
-
-[sub_resource type="CompressedTexture2D" id="CompressedTexture2D_yltcx"]
+[ext_resource type="Texture2D" uid="uid://bt0eyuhxpem4u" path="res://Cycle/UI/clock_hand.png" id="3_ewqxd"]
 
 [node name="DayNightCycleUI" type="Control"]
 texture_filter = 1
@@ -11,6 +11,7 @@ layout_mode = 3
 anchors_preset = 0
 offset_right = 149.0
 offset_bottom = 86.0
+script = ExtResource("1_jcn3g")
 
 [node name="TextureRect" type="TextureRect" parent="."]
 layout_mode = 0
@@ -72,7 +73,9 @@ text = "12:00 am"
 [node name="Arrow" type="TextureRect" parent="."]
 unique_name_in_owner = true
 layout_mode = 0
-offset_top = 23.0
-offset_right = 40.0
-offset_bottom = 40.0
-texture = SubResource("CompressedTexture2D_yltcx")
+offset_left = 99.0
+offset_top = 36.0
+offset_right = 139.0
+offset_bottom = 53.0
+pivot_offset = Vector2(5, 8.5)
+texture = ExtResource("3_ewqxd")
